requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Django==5.1.7
djangorestframework==3.15.2
idna==3.10
libretranslatepy==2.1.1
lxml==5.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
translate==3.6.1
urllib3==2.3.0
